♻️ Change bus width to usize

+++ src/common.rs
@@ -1,1 +1,1 @@
-pub const BUS_WIDTH: u8 = 8;
+pub const BUS_WIDTH: usize = 8;
